pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-transformer"
version = "0.1.0"
description = "Deterministic prompt transformation service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.115.0",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.4.0",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.34",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
